V-02/F-03 Рефакторинг кода
