+++ src/components/OrderPage/OrderPage.js
@@ -4,9 +4,8 @@ import { useParams } from "react-router-dom";
 import AsideMenu from "../common/AsideMenu/AsideMenu";
 import TopMenu from "../common/TopMenu/TopMenu";
 import Breadcrumbs from "./Breadcrumbs/Breadcrumbs";
-import ModelSelection from "./ModelSelection/ModelSelection";
 import OrderDetailsContainer from "./OrderDetails/OrderDetailsContainer";
-import PointSelectionContainer from "./PointSelection/PointSelectionContainer";
+import PointSelection from "./PointSelection/PointSelection";
 import "./styles.scss";
 
 const OrderPage = () => {
@@ -21,11 +20,13 @@ const OrderPage = () => {
   const getCurrentStepContent = () => {
     switch (currentStep) {
       case "point":
-        return <PointSelectionContainer />;
+        return <PointSelection />;
       default:
         return null;
     }
   };
+
+  const сurrentStepContent = getCurrentStepContent();
 
   return (
     <div className="order-page">
@@ -43,9 +44,7 @@ const OrderPage = () => {
             </div>
             <div className="order-container">
               <div className="order-container__content content main-wrapper">
-                <div className="content__left-side">
-                  {getCurrentStepContent()}
-                </div>
+                <div className="content__left-side">{сurrentStepContent}</div>
                 <div className="content__vertical-line" />
                 <div className="content__right-side">
                   <OrderDetailsContainer />
