+++ src/components/common/Autocomplete/styles.scss
@@ -1,0 +1,41 @@
+@import "../../../scss/variables.scss";
+
+/* Autocomplete */
+
+.autocomplete {
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  max-height: 106px;
+  width: 171px;
+  background-color: $white;
+  border: 1px solid $gray-light;
+  z-index: 3;
+  overflow-y: auto;
+
+  .autocomplete__item {
+    padding: 8px;
+
+    &:hover {
+      color: $main-accent;
+    }
+  }
+
+  &::-webkit-scrollbar {
+    width: 8px;
+    background-color: $white;
+    border-left: 1px solid $gray-light;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    width: 2px;
+    background-color: $gray;
+    border-radius: 4px;
+    border: 2px solid transparent;
+    background-clip: content-box;
+  }
+
+  @media (min-width: $tablet-width) {
+    width: 222px;
+  }
+}
